Rename test project and namespace
