+++ Medidata.ZipkinTracerModule.sln
@@ -13,8 +13,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{ACDBD6
 		.nuget\packages.config = .nuget\packages.config
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Medidata.ZipkinTracer.Core.Test", "tests\Medidata.ZipkinTracer.Core.Test\Medidata.ZipkinTracer.Core.Test.csproj", "{0158C8BD-D881-4EE8-A8C6-54707A67A00D}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{7CF24590-AC06-44E4-AF93-88A9B2CD1264}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
@@ -23,6 +21,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Medidata.ZipkinTracer.Owin", "src\Medidata.ZipkinTracer.Owin\Medidata.ZipkinTracer.Owin.csproj", "{86DFF8D6-DFEA-4D7F-9E57-0709397059B7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Medidata.ZipkinTracer.WebApi", "src\Medidata.ZipkinTracer.WebApi\Medidata.ZipkinTracer.WebApi.csproj", "{C552DC79-0419-4FDA-908C-6D260F41FD4C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Medidata.ZipkinTracer.Tests", "tests\Medidata.ZipkinTracer.Tests\Medidata.ZipkinTracer.Tests.csproj", "{0158C8BD-D881-4EE8-A8C6-54707A67A00D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -34,10 +34,6 @@ Global
 		{EB432891-204B-4B97-BFB1-A820E424284E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EB432891-204B-4B97-BFB1-A820E424284E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EB432891-204B-4B97-BFB1-A820E424284E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0158C8BD-D881-4EE8-A8C6-54707A67A00D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0158C8BD-D881-4EE8-A8C6-54707A67A00D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0158C8BD-D881-4EE8-A8C6-54707A67A00D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0158C8BD-D881-4EE8-A8C6-54707A67A00D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{86DFF8D6-DFEA-4D7F-9E57-0709397059B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{86DFF8D6-DFEA-4D7F-9E57-0709397059B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{86DFF8D6-DFEA-4D7F-9E57-0709397059B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -46,6 +42,10 @@ Global
 		{C552DC79-0419-4FDA-908C-6D260F41FD4C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C552DC79-0419-4FDA-908C-6D260F41FD4C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C552DC79-0419-4FDA-908C-6D260F41FD4C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0158C8BD-D881-4EE8-A8C6-54707A67A00D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0158C8BD-D881-4EE8-A8C6-54707A67A00D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0158C8BD-D881-4EE8-A8C6-54707A67A00D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0158C8BD-D881-4EE8-A8C6-54707A67A00D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
