+++ src/Medidata.ZipkinTracer.Core/Properties/AssemblyInfo.cs
@@ -22,4 +22,4 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyVersion("4.0.0")]
 [assembly: AssemblyFileVersion("4.0.0")]
 [assembly: AssemblyInformationalVersion("4.0.0-preview0001")]
-[assembly: InternalsVisibleTo("Medidata.ZipkinTracer.Core.Test")]
+[assembly: InternalsVisibleTo("Medidata.ZipkinTracer.Tests")]
