+++ tests/Medidata.ZipkinTracer.Tests/Properties/AssemblyInfo.cs
@@ -1,0 +1,15 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("Medidata.ZipkinTracer.Tests")]
+[assembly: AssemblyDescription("Test assembly for the Zipkin Tracer Module.")]
+[assembly: AssemblyCompany("Medidata Solutions, Inc.")]
+[assembly: AssemblyProduct("Medidata.ZipkinTracerModule")]
+[assembly: AssemblyCopyright("Copyright © Medidata Solutions Inc., 2017")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+[assembly: ComVisible(false)]
+[assembly: Guid("78961bfd-538b-4a56-8d2d-7cd6b7e6e65d")]
+[assembly: AssemblyVersion("4.0.0")]
+[assembly: AssemblyFileVersion("4.0.0")]
+[assembly: AssemblyInformationalVersion("4.0.0-preview0001")]
