+++ tests/Medidata.ZipkinTracer.Tests/SpanProcessorTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Concurrent;
+using Medidata.ZipkinTracer.Core;
 using Medidata.ZipkinTracer.Core.Logging;
 using Medidata.ZipkinTracer.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -7,7 +8,7 @@ using Ploeh.AutoFixture;
 using Rhino.Mocks;
 using Rhino.Mocks.Interfaces;
 
-namespace Medidata.ZipkinTracer.Core.Collector.Test
+namespace Medidata.ZipkinTracer.Tests
 {
     [TestClass]
     public class SpanProcessorTests
