+++ tests/Medidata.ZipkinTracer.Tests/ZipkinConfigTests.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
+using Medidata.ZipkinTracer.Core;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Ploeh.AutoFixture;
 
-namespace Medidata.ZipkinTracer.Core.Test
+namespace Medidata.ZipkinTracer.Tests
 {
     [TestClass]
     public class ZipkinConfigTests
